% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_jd.R
\name{kna_getDate_jd}
\alias{kna_getDate_jd}
\title{读取新通路的标准化后的日期}
\usage{
kna_getDate_jd(dir_name = "data-raw/jlnka_202011/jd_202011_downloads")
}
\arguments{
\item{dir_name}{目录}
}
\value{
返回日期字段
}
\description{
读取新通路的标准化后的日期
}
\examples{
kna_getDate_jd()
}
