% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_watsons.R
\name{nka_getCategory_watsons}
\alias{nka_getCategory_watsons}
\title{获取watson的类型}
\usage{
nka_getCategory_watsons(
  dir_name = "data-raw/jlnka_202011/Watsons_202011_downloads/xlsx"
)
}
\arguments{
\item{dir_name}{目录}
}
\value{
返回值
}
\description{
获取watson的类型
}
\examples{
nka_getCategory_watsons()
}
