% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_rtm.R
\name{nka_getUserId_rtm}
\alias{nka_getUserId_rtm}
\title{获取用户账号}
\usage{
nka_getUserId_rtm(
  dir_name = "data-raw/jlnka_202011/rtm_202011_downloads/Daily"
)
}
\arguments{
\item{dir_name}{目录名}
}
\value{
返回值
}
\description{
获取用户账号
}
\examples{
kna_rtm_getUserId()
}
