% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_carrefour.R
\name{nka_readSrc_carrefour}
\alias{nka_readSrc_carrefour}
\title{读取家乐福数据}
\usage{
nka_readSrc_carrefour(
  file = "data-raw/jlnka_202011/carrefour_202011_downloads/carrefour_202011.xlsx",
  file_short = "carrefour_202011.xlsx",
  FUser = "RDS"
)
}
\arguments{
\item{file}{文件名}

\item{file_short}{短文件名}

\item{FUser}{用户}
}
\value{
返回值
}
\description{
读取家乐福数据
}
\examples{
nka_readSrc_carrefour()
}
