% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_jd.R
\name{nka_readSrc_jd}
\alias{nka_readSrc_jd}
\title{新通路读取数据源}
\usage{
nka_readSrc_jd(
  file = "data-raw/jlnka_202011/jd_202011_downloads/JD_20201101.xlsx",
  file_short = "JD_20201101.xlsx",
  FSaleDate = "2020-11-01",
  FUser = "RDS"
)
}
\arguments{
\item{file}{文件}

\item{file_short}{短文件}

\item{FSaleDate}{销售日期}

\item{FUser}{用户}
}
\value{
返回表
}
\description{
新通路读取数据源
}
\examples{
nka_readSrc_jd()
}
