% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_rtm.R
\name{nka_readSrc_rtm_daily}
\alias{nka_readSrc_rtm_daily}
\title{计取日报数据}
\usage{
nka_readSrc_rtm_daily(
  file = "data-raw/jlnka_202011/rtm_202011_downloads/Daily/RTM_rt210602_2020-11-01.xlsx",
  file_short = "RTM_rt210602_2020-11-01.xlsx",
  FLoginAcct = "rt210602",
  FUser = "RDS"
)
}
\arguments{
\item{file}{文件}

\item{file_short}{短文件}

\item{FLoginAcct}{登录账号}

\item{FUser}{用户}
}
\value{
返回值
}
\description{
计取日报数据
}
\examples{
nka_readSrc_rtm_daily()
}
