% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_walmart.R
\name{nka_readSrc_walmart}
\alias{nka_readSrc_walmart}
\title{读取walmart数据源}
\usage{
nka_readSrc_walmart(
  file = "data-raw/walmart_202011_downloads/xlsx/walmart_202011_01.xlsx",
  FUser = "RDS"
)
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
读取walmart数据源
}
\examples{
nka_readSrc_walmart()
}
