% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_watsons.R
\name{nka_readSrc_watsons}
\alias{nka_readSrc_watsons}
\title{读取数据}
\usage{
nka_readSrc_watsons(
  file = "data-raw/jlnka_202011/Watsons_202011_downloads/xlsx/CX_20201102.xlsx",
  file_short = "CX_20201102.xlsx",
  FCategory = "春夏",
  FUser = "RDS"
)
}
\arguments{
\item{file}{文件名}

\item{file_short}{短文件名}

\item{FCategory}{分类}

\item{FUser}{用户}
}
\value{
返回值
}
\description{
读取数据
}
\examples{
nka_readSrc_watsons()
}
