% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_carrefour.R
\name{nka_writeSrc_carrefour}
\alias{nka_writeSrc_carrefour}
\title{写入carrefour数据}
\usage{
nka_writeSrc_carrefour(
  file = "data-raw/jlnka_202011/carrefour_202011_downloads/carrefour_202011.xlsx",
  file_short = "carrefour_202011.xlsx",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_carrefour"
)
}
\arguments{
\item{file}{文件名}

\item{file_short}{短文件名}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
写入carrefour数据
}
\examples{
nka_writeSrc_carrefour()
}
