% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_jd.R
\name{nka_writeSrc_jd}
\alias{nka_writeSrc_jd}
\title{写入京东数据}
\usage{
nka_writeSrc_jd(
  file = "data-raw/jlnka_202011/jd_202011_downloads/JD_20201101.xlsx",
  file_short = "JD_20201101.xlsx",
  FSaleDate = "2020-11-01",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_jd"
)
}
\arguments{
\item{file}{文件}

\item{file_short}{短文件}

\item{FSaleDate}{销售日期}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
写入京东数据
}
\examples{
nka_writeSrc_jd()
}
