% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_rtm.R
\name{nka_writeSrc_rtm_daily}
\alias{nka_writeSrc_rtm_daily}
\title{写入大润发日数据}
\usage{
nka_writeSrc_rtm_daily(
  file = "data-raw/jlnka_202011/rtm_202011_downloads/Daily/RTM_rt210602_2020-11-01.xlsx",
  file_short = "RTM_rt210602_2020-11-01.xlsx",
  FLoginAcct = "rt210602",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_rtm_day"
)
}
\arguments{
\item{file}{文件名}

\item{file_short}{短文件名}

\item{FLoginAcct}{walmart账号}

\item{FUser}{写入中间表用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
写入大润发日数据
}
\examples{
nka_writeSrc_rtm_daily()
}
