% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_rtm.R
\name{nka_writeSrc_rtm_monthly}
\alias{nka_writeSrc_rtm_monthly}
\title{写入月报数据}
\usage{
nka_writeSrc_rtm_monthly(
  file = "data-raw/jlnka_202011/rtm_202011_downloads/Month/RTM_rt211571_2020-11.xlsx",
  file_short = "RTM_rt211571_2020-11.xlsx",
  FLoginAcct = "rt211571",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_rtm_month"
)
}
\arguments{
\item{file}{文件}

\item{file_short}{短文件名}

\item{FLoginAcct}{walmart登录名}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
写入月报数据
}
\examples{
nka_writeSrc_rtm_monthly()
}
