% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_walmart.R
\name{nka_writeSrc_walmart}
\alias{nka_writeSrc_walmart}
\title{写入数据源}
\usage{
nka_writeSrc_walmart(
  file = "data-raw/walmart_202011_downloads/xlsx/walmart_202011_01.xlsx",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_walmart"
)
}
\arguments{
\item{file}{文件}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
写入结果
}
\description{
写入数据源
}
\examples{
nka_writeSrc_walmart()
}
