% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_watsons.R
\name{nka_writeSrc_watsons}
\alias{nka_writeSrc_watsons}
\title{写入watson的数据源}
\usage{
nka_writeSrc_watsons(
  file = "data-raw/jlnka_202011/Watsons_202011_downloads/xlsx/CX_20201102.xlsx",
  file_short = "CX_20201102.xlsx",
  FCategory = "春夏",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_watsons"
)
}
\arguments{
\item{file}{文件名}

\item{file_short}{短文件名}

\item{FCategory}{类别}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
写入watson的数据源
}
\examples{
nka_writeSrc_watsons()
}
