% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_jd.R
\name{nka_writeSrcBatch_jd}
\alias{nka_writeSrcBatch_jd}
\title{批量写入京东数据}
\usage{
nka_writeSrcBatch_jd(
  dir_name = "data-raw/jlnka_202011/jd_202011_downloads",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_jd"
)
}
\arguments{
\item{dir_name}{目录}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
批量写入京东数据
}
\examples{
nka_writeSrcBatch_jd()
}
