% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_rtm.R
\name{nka_writeSrcBatch_rtm_daily}
\alias{nka_writeSrcBatch_rtm_daily}
\title{批量写入日报数据}
\usage{
nka_writeSrcBatch_rtm_daily(
  dir_name = "data-raw/jlnka_202011/rtm_202011_downloads/Daily",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_rtm_day"
)
}
\arguments{
\item{dir_name}{目录}

\item{FUser}{写入中间表用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
批量写入日报数据
}
\examples{
nka_writeSrcBatch_rtm_daily()
}
