% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_rtm.R
\name{nka_writeSrcBatch_rtm_monthly}
\alias{nka_writeSrcBatch_rtm_monthly}
\title{批量写入月报数据}
\usage{
nka_writeSrcBatch_rtm_monthly(
  dir_name = "data-raw/jlnka_202011/rtm_202011_downloads/Month",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_rtm_month"
)
}
\arguments{
\item{dir_name}{目标}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
批量写入月报数据
}
\examples{
nka_writeSrcBatch_rtm_monthly()
}
