% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_walmart.R
\name{nka_writeSrcBatch_walmart}
\alias{nka_writeSrcBatch_walmart}
\title{批量写入walmart数据源}
\usage{
nka_writeSrcBatch_walmart(
  dir = "data-raw/walmart_202011_downloads/xlsx",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_walmart"
)
}
\arguments{
\item{dir}{路径}

\item{FUser}{用户}

\item{conn}{连接}

\item{table_name}{表名}
}
\value{
无返回值
}
\description{
批量写入walmart数据源
}
\examples{
nka_writeSrcBatch_walmart()
}
