% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_watsons.R
\name{nka_writeSrcBatch_watsons}
\alias{nka_writeSrcBatch_watsons}
\title{批量写入数据}
\usage{
nka_writeSrcBatch_watsons(
  dir_name = "data-raw/jlnka_202011/Watsons_202011_downloads/xlsx",
  FUser = "RDS",
  conn = sqlr::sql_conn_common(db_name = "jlnka"),
  table_name = "t_nka_src_watsons"
)
}
\arguments{
\item{dir_name}{目录名称}

\item{FUser}{用户}

\item{conn}{连接名}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
批量写入数据
}
\examples{
nka_writeSrcBatch_watsons()
}
